backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
uvicorn==0.24.0
googlemaps==4.10.0
scikit-learn==1.3.2
numpy==1.26.2
python-multipart>=0.0.5
python-dotenv==1.0.0

